# shafurui-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "shafurui-backend HTTP API documentation.",
        "title": "shafurui-backend API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:6305",
    "basePath": "/api",
    "paths": {
        "/app/getHelloInfo": {
            "post": {
                "description": "根据传入 name 返回问候信息。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app"
                ],
                "summary": "获取问候信息",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.GetHelloInfoRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "description": "使用管理员用户名和密码换取 accessToken 与 refreshToken。业务失败同样返回 HTTP 200，通过 code/message 区分。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "管理员登录",
                "parameters": [
                    {
                        "description": "登录参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AuthLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.SwaggerAuthLoginResponse"
                        }
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "description": "使用 refreshToken 换取新的 accessToken 与 refreshToken。业务失败同样返回 HTTP 200，通过 code/message 区分。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "刷新管理员 Token",
                "parameters": [
                    {
                        "description": "刷新参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AuthRefreshTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.SwaggerAuthLoginResponse"
                        }
                    }
                }
            }
        },
        "/meta": {
            "get": {
                "description": "返回 public/meta.json 中的动态 JSON 内容；文件缺失时返回 deployTime=unknown。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meta"
                ],
                "summary": "获取部署元信息",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.MyCode": {
            "type": "integer",
            "format": "int64",
            "enum": [
                1000,
                1001,
                1002,
                1003,
                1004,
                1005,
                1006,
                1007,
                1008,
                1009,
                1010
            ],
            "x-enum-comments": {
                "CodeInvalidParam": "请求参数错误",
                "CodeInvalidPassword": "用户名或密码错误",
                "CodeInvalidToken": "无效的token",
                "CodeNeedLogin": "需要登录",
                "CodePermissionDenied": "权限不足",
                "CodeResourceExists": "资源已存在",
                "CodeResourceNotExist": "资源不存在",
                "CodeServerBusy": "服务繁忙",
                "CodeSuccess": "成功",
                "CodeUserExist": "用户已存在",
                "CodeUserNotExist": "用户不存在"
            },
            "x-enum-descriptions": [
                "成功",
                "请求参数错误",
                "用户已存在",
                "用户不存在",
                "用户名或密码错误",
                "服务繁忙",
                "需要登录",
                "无效的token",
                "资源已存在",
                "资源不存在",
                "权限不足"
            ],
            "x-enum-varnames": [
                "CodeSuccess",
                "CodeInvalidParam",
                "CodeUserExist",
                "CodeUserNotExist",
                "CodeInvalidPassword",
                "CodeServerBusy",
                "CodeNeedLogin",
                "CodeInvalidToken",
                "CodeResourceExists",
                "CodeResourceNotExist",
                "CodePermissionDenied"
            ]
        },
        "controller.SwaggerAuthLoginResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/controller.MyCode"
                        }
                    ],
                    "example": 1000
                },
                "data": {
                    "$ref": "#/definitions/model.AuthLoginResponse"
                },
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "controller.SwaggerErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/controller.MyCode"
                        }
                    ],
                    "example": 1001
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "请求参数错误"
                }
            }
        },
        "controller.SwaggerHelloInfoResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/controller.MyCode"
                        }
                    ],
                    "example": 1000
                },
                "data": {
                    "$ref": "#/definitions/model.GetHelloInfoResponse"
                },
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "model.AuthLoginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "model.AuthLoginResponse": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string"
                },
                "refreshToken": {
                    "type": "string"
                }
            }
        },
        "model.AuthRefreshTokenRequest": {
            "type": "object",
            "required": [
                "refreshToken"
            ],
            "properties": {
                "refreshToken": {
                    "type": "string"
                }
            }
        },
        "model.GetHelloInfoRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "model.GetHelloInfoResponse": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}